backend/requirements.txt
openai
openai-agents
requests
python-dotenv
anthropic
fastapi
uvicorn
pydantic
asyncio